// src/modules/SingleCompany/types/IReviewForm.types.ts
export interface IReviewForm {
    benefits: string[];
    company_id: string;
    cons: string;
    experience_duration: string;
    is_current_employee: boolean;
    position: string;
    pros: string;
    rating_growth_opportunities: number;
    rating_management: number;
    rating_relax_conditions: number;
    rating_salary_level: number;
    rating_team: number;
    rating_work_conditions: number;
    region: string;
    user_id: string;
}